move crypto utils
